+++ CHANGES.rst
@@ -1,8 +1,15 @@
 Changelog
 =========
 
-14.1 (unreleased)
+15.0 (unreleased)
 -----------------
+
+Breaking changes
+++++++++++++++++
+
+- Drop support for Python 3.8.
+
+- Drop support for pytest < 7.4.
 
 Features
 ++++++++
@@ -10,6 +17,7 @@ Features
 - Fix compatibility with pytest 8.2.
   (`#267 <https://github.com/pytest-dev/pytest-rerunfailures/issues/267>`_)
 
+- Add support for pytest 8.2, 8.3.
 
 - Add ``--fail-on-flaky`` option to fail the test run with custom exit code
   when test passed on rerun.